feat: add coordination shutdown mechanism

+++ internal/osutil/signal.go
@@ -27,6 +27,7 @@ package osutil
 import (
 	"os"
 	"os/signal"
+	"runtime"
 	"sync"
 	"syscall"
 
@@ -70,11 +71,13 @@ func HandleSignals(logger log.Logger, cancel <-chan types.Unit) {
 			}
 
 			signal.Stop(notifier)
-			osid := syscall.Getpid()
-			if osid == 1 {
-				os.Exit(0)
+			if runtime.GOOS != "windows" {
+				osid := syscall.Getpid()
+				if osid == 1 {
+					os.Exit(0)
+				}
+				_ = syscall.Kill(osid, sig.(syscall.Signal))
 			}
-			_ = syscall.Kill(osid, sig.(syscall.Signal))
 		case <-cancel:
 			return
 		}
